// apps/onestock-extensions/extensions/pickup-point-delivery-option-generators/generated/api.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /**
   * Represents an [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601)-encoded date string.
   * For example, September 7, 2019 is represented as `"2019-07-16"`.
   */
  Date: { input: any; output: any; }
  /**
   * Represents an [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601)-encoded date and time string.
   * For example, 3:50 pm on September 7, 2019 in the time zone of UTC (Coordinated Universal Time) is
   * represented as `"2019-09-07T15:50:00Z`".
   */
  DateTime: { input: any; output: any; }
  /**
   * A subset of the [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) format that
   * includes the date and time but not the timezone which is determined from context.
   *
   * For example, "2018-01-01T00:00:00".
   */
  DateTimeWithoutTimezone: { input: any; output: any; }
  /**
   * A signed decimal number, which supports arbitrary precision and is serialized as a string.
   *
   * Example values: `"29.99"`, `"29.999"`.
   */
  Decimal: { input: any; output: any; }
  /**
   * A function-scoped handle to a refer a resource.
   * The Handle type appears in a JSON response as a String, but it is not intended to be human-readable.
   * Example value: `"10079785100"`
   */
  Handle: { input: any; output: any; }
  /**
   * A [JSON](https://www.json.org/json-en.html) object.
   *
   * Example value:
   * `{
   *   "product": {
   *     "id": "gid://shopify/Product/1346443542550",
   *     "title": "White T-shirt",
   *     "options": [{
   *       "name": "Size",
   *       "values": ["M", "L"]
   *     }]
   *   }
   * }`
   */
  JSON: { input: any; output: any; }
  /**
   * A subset of the [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) format that
   * includes the time but not the date or timezone which is determined from context.
   * For example, "05:43:21".
   */
  TimeWithoutTimezone: { input: any; output: any; }
  /**
   * Represents an [RFC 3986](https://datatracker.ietf.org/doc/html/rfc3986) and
   * [RFC 3987](https://datatracker.ietf.org/doc/html/rfc3987)-compliant URI string.
   *
   * For example, `"https://example.myshopify.com"` is a valid URL. It includes a scheme (`https`) and a host
   * (`example.myshopify.com`).
   */
  URL: { input: any; output: any; }
  /** A void type that can be used to return a null value from a mutation. */
  Void: { input: any; output: any; }
};

/** Information about the allocation for one or more line items that are to be delivered to a specific address. */
export type Allocation = {
  __typename?: 'Allocation';
  /** The delivery address for the allocation. */
  deliveryAddress: MailingAddress;
};
// ...
export type BusinessHours = {
  /** The day of the week. */
  day: Weekday;
  /**
   * The operating time periods for a specific day. To represent split schedules
   * (for example, 9:00-12:00 and 13:00-17:00 for locations that close during
   * lunch), you can specify multiple periods.
   *
   * Each period consists of `opening_time` and `closing_time`. To indicate the
   * location is closed on a specific day, such as Sunday, then omit it entirely
   * from the `business_hours` array.
   */
  periods: Array<BusinessHoursPeriod>;
};

/** The pickup point's hours of operation for a specific day of the week. */
export type BusinessHoursPeriod = {
  /** The time that the pickup point location closes. */
  closingTime: Scalars['TimeWithoutTimezone']['input'];
  /** The time that the pickup point location opens. */
  openingTime: Scalars['TimeWithoutTimezone']['input'];
};
// ...
export type FunctionRunResult = {
  /** An ordered list of operations to apply to generate pickup point delivery options. */
  operations: Array<Operation>;
};
// ...
export type HttpResponse = {
  __typename?: 'HttpResponse';
  /**
   * The HTTP response body as a plain string.
   * Use this field when the body is not in JSON format.
   */
  body?: Maybe<Scalars['String']['output']>;
  /** An HTTP header. */
  header?: Maybe<HttpResponseHeader>;
  /**
   * The HTTP headers.
   * @deprecated Use `header` instead.
   */
  headers: Array<HttpResponseHeader>;
  /**
   * The HTTP response body parsed as JSON.
   * If the body is valid JSON, it will be parsed and returned as a JSON object.
   * If parsing fails, then raw body is returned as a string.
   * Use this field when you expect the response to be JSON, or when you're dealing
   * with mixed response types, meaning both JSON and non-JSON.
   * Using this field reduces function instruction consumption and ensures that the data is formatted in logs.
   * To prevent increasing the function target input size unnecessarily, avoid querying
   * both `body` and `jsonBody` simultaneously.
   */
  jsonBody?: Maybe<Scalars['JSON']['output']>;
  /** The HTTP status code. */
  status: Scalars['Int']['output'];
};


/** The attributes associated with an HTTP response. */
export type HttpResponseHeaderArgs = {
  name: Scalars['String']['input'];
};

/** The attributes associated with an HTTP response header. */
export type HttpResponseHeader = {
  __typename?: 'HttpResponseHeader';
  /** Header name. */
  name: Scalars['String']['output'];
  /** Header value. */
  value: Scalars['String']['output'];
};

export type Input = {
  __typename?: 'Input';
  /**
   * A list of one or more line items that are to be delivered to a specific address.
   * @deprecated Use `deliveryAddress` instead.
   */
  allocations: Array<Allocation>;
  /**
   * The cart where the Function is running. A cart contains the merchandise that a customer intends to purchase
   * and information about the customer, such as the customer's email address and phone number.
   */
  cart: Cart;
  /** The delivery address for the pickup point delivery option. */
  deliveryAddress: MailingAddress;
  /** The fetch target result. Your Function must return this data structure when generating the request. */
  fetchResult?: Maybe<HttpResponse>;
  /**
   * The regional and language settings that determine how the Function
   * handles currency, numbers, dates, and other locale-specific values
   * during discount calculations. These settings are based on the store's configured
   * [localization practices](https://shopify.dev/docs/apps/build/functions/localization-practices-shopify-functions).
   */
  localization: Localization;
  /**
   * The exchange rate used to convert discounts between the shop's default
   * currency and the currency that displays to the customer during checkout.
   * For example, if a store operates in USD but a customer is viewing discounts in EUR,
   * then the presentment currency rate handles this conversion for accurate pricing.
   */
  presentmentCurrencyRate: Scalars['Decimal']['output'];
  /**
   * Information about the shop where the Function is running, including the shop's timezone
   * setting and associated [metafields](https://shopify.dev/docs/apps/build/custom-data).
   */
  shop: Shop;
};
// ...
export type Localization = {
  __typename?: 'Localization';
  /**
   * The country for which the store is customized, reflecting local preferences and regulations.
   * Localization might influence the language, currency, and product offerings available in a store to enhance
   * the shopping experience for customers in that region.
   */
  country: Country;
  /**
   * The language for which the store is customized, ensuring content is tailored to local customers.
   * This includes product descriptions and customer communications that resonate with the target audience.
   */
  language: Language;
  /**
   * The market of the active localized experience.
   * @deprecated This `market` field will be removed in a future version of the API.
   */
  market: Market;
};
// ...
export type MetafieldOutput = {
  /**
   * The unique identifier for the metafield within its namespace. A metafield is composed of a
   * namespace and a key, in the format `namespace.key`.
   */
  key: Scalars['String']['input'];
  /**
   * A category that organizes a group of metafields. Namespaces are used to prevent naming conflicts
   * between different apps or different parts of the same app. If omitted, then the
   * [app-reserved namespace](https://shopify.dev/docs/apps/build/custom-data/ownership)
   * is used.
   */
  namespace?: InputMaybe<Scalars['String']['input']>;
  /**
   * The [type of data](https://shopify.dev/docs/apps/build/custom-data/metafields/list-of-data-types)
   * that's stored in the metafield.
   */
  type: Scalars['String']['input'];
  /** The data to store in the metafield. */
  value: Scalars['String']['input'];
};
// ...
export type Operation = {
  /** The pickup point delivery option to add. */
  add: PickupPointDeliveryOption;
};
// ...
export type PickupAddress = {
  /** Address line 1. */
  address1: Scalars['String']['input'];
  /** Address line 2. */
  address2?: InputMaybe<Scalars['String']['input']>;
  /** City. */
  city: Scalars['String']['input'];
  /** Country. */
  country?: InputMaybe<Scalars['String']['input']>;
  /** Country code. */
  countryCode: CountryCode;
  /** Latitude. */
  latitude: Scalars['Float']['input'];
  /** Longitude. */
  longitude: Scalars['Float']['input'];
  /** Phone number. */
  phone?: InputMaybe<Scalars['String']['input']>;
  /** Province. */
  province?: InputMaybe<Scalars['String']['input']>;
  /** Province code. */
  provinceCode?: InputMaybe<Scalars['String']['input']>;
  /** Zip code. */
  zip?: InputMaybe<Scalars['String']['input']>;
};

/**
 * A third-party location where customers can collect their orders, such as a local
 * post office, convenience store, or self-service locker.
 */
export type PickupPoint = {
  /** The address of the pickup point's location. */
  address: PickupAddress;
  /**
   * A list of operating hours for the pickup point. A day that's omitted or
   * doesn't have defined time periods is considered to be closed.
   */
  businessHours?: InputMaybe<Array<BusinessHours>>;
  /** The unique ID that the third-party service has assigned to the pickup point. */
  externalId: Scalars['String']['input'];
  /** The name that the third-party service has assigned to the pickup point. */
  name: Scalars['String']['input'];
  /** The third-party service that manages pickup point delivery options. */
  provider: Provider;
};

/** A pickup point delivery option, such as a local post office, convenience store, or self-service locker. */
export type PickupPointDeliveryOption = {
  /**
   * The delivery option's cost in the cart's currency. If not provided, then the
   * system uses the default price from the location's pickup points settings.
   */
  cost?: InputMaybe<Scalars['Decimal']['input']>;
  /** The metafields associated with the pickup point delivery option. */
  metafields?: InputMaybe<Array<MetafieldOutput>>;
  /**
   * The [pickup point](https://help.shopify.com/manual/fulfillment/shopify-shipping/pickup-points) where customers can collect their online orders rather than having them
   * delivered to their home or to a [local pickup point](https://help.shopify.com/manual/fulfillment/setup/delivery-methods/pickup-in-store).
   * Examples of pickup points include local post offices, convenience stores, and
   * self-service lockers.
   */
  pickupPoint: PickupPoint;
};
// ...
export type Provider = {
  /**
   * The URL logo for the third-party service that manages the pickup point
   * delivery options. The base URL must be `https://cdn.shopify.com`.
   */
  logoUrl: Scalars['URL']['input'];
  /** The name of the third-party service that manages the pickup point delivery options. */
  name: Scalars['String']['input'];
};
// ...
export enum Weekday {
  /** Friday */
  Friday = 'FRIDAY',
  /** Monday */
  Monday = 'MONDAY',
  /** Saturday */
  Saturday = 'SATURDAY',
  /** Sunday */
  Sunday = 'SUNDAY',
  /** Thursday */
  Thursday = 'THURSDAY',
  /** Tuesday */
  Tuesday = 'TUESDAY',
  /** Wednesday */
  Wednesday = 'WEDNESDAY'
}
// ...
export type RunInput = { __typename?: 'Input', fetchResult?: { __typename?: 'HttpResponse', status: number, body?: string | null } | null };